Add some initial sketches of ARM details
There are 37 total registers, per documentation. Not all of them are
available at any point in time.

I have the first handful of thumb instructions sketched out -- just the
type declarations, no encoding or decoding of opcodes at this point. I'm
out of time tonight but will pick up the rest of those when I get a
chance. Modeling the instructions in Zig is even easier than I expected
it to be.