Minor adjustments to Pipeline
This wasn't much progress yet. I'm still feeling out how everything
lines up in that component.

I don't want to have to allocate for the Core, but without allocating,
we cannot safely take the reference to a register that the core owns.
That is making handling of the pc value awkward now.