Clean up Pipeline type
This has a couple key differences which I think are more idiomatic Zig
patterns than what the file previously had.

First, the union types that back the fields of the Pipeline struct, are
now defined within the Pipeline struct, so they are not scoped at the
same visibility exactly.

Next, we stop using optionals for those enum field values. Using
optionals seemed to make sense at first, but then handling the default
absent case became more code than the present cases, which was a hint
that it made more sense to include, in this case, "idle" as a possible
state for those values.

Last, we use the file as the struct to avoid declaring a redundant
Pipeline type within pipeline.zig.