Sketch out MemoryManager and integrate with Core
Now be more careful to use CamelCase filenames for files with top-level
structs.

This is the general structure I had in mind, where the pipeline, CPU,
memory manager, and so on are able to respond to ticks independent of
one another, and then the Core handles feedback between individual
components. I think it will be a while before we are able to pin down
any not-yet-realized timing issues, but adjustments should be
straightforward in this architecture.