Add MemoryManager type
Which is constituted of many MemoryRegion values, and it manages pending
reads and writes from various components (just the pipeline and CPU so
far, but eventually DMA too).

There is a bit of cleanup to do yet here, but that will be easier to do
as separate small commits now.